requirements.txt
Django>=4.2,<5.1
wagtail>=6.2,<6.3
psycopg2-binary
gdal
django-sass
dj-database-url
pillow
python-dateutil
fiona
pyproj
geos
shapely
rasterio
django-auditlog
django-autoslug
django-sass
django-json-widget
django-guardian
django-allauth
django-polymorphic
django-modelcluster
djoser
djangorestframework
djangorestframework-gis
django-filter
earthengine-api
httpx
python-dotenv[cli]
gunicorn
django-organizations
django-extensions
django-countries
django-modeltranslation
ipython
redis
django-axes[ipware]
whitenoise
drf-spectacular
django-session-timeout
django-shortuuidfield
django_auditlog
google-cloud-storage
google-cloud-recaptcha-enterprise
google-cloud
django-storages
django-cors-headers
dj-database-url
firebase_admin
django-jazzmin
